+++ truffle/src/com.oracle.truffle.api.dsl.test/src/com/oracle/truffle/api/dsl/test/examples/Interop.java
@@ -71,31 +71,31 @@ public class Interop {
         assertEquals(42, target.call(o1, 42));
         assertEquals(43, target.call(o2, 43));
         assertEquals(44, target.call(o3, 44));
-        assertEquals(3, node.cached);
-        assertEquals(0, node.generic);
+        assertEquals(3, node.cachedCount);
+        assertEquals(0, node.genericCount);
         assertEquals(45, target.call(o4, 45)); // operation gets generic
         assertEquals(42, target.call(o1, 42));
         assertEquals(43, target.call(o2, 43));
         assertEquals(44, target.call(o3, 44));
-        assertEquals(3, node.cached);
-        assertEquals(4, node.generic);
+        assertEquals(3, node.cachedCount);
+        assertEquals(4, node.genericCount);
     }
 
     public static class UseInterop extends ExampleNode {
 
-        int cached = 0;
-        int generic = 0;
+        int cachedCount = 0;
+        int genericCount = 0;
 
         @Specialization(guards = "operation.accept(target)")
         protected Object interopCached(VirtualFrame frame, TruffleObject target, Object value, //
                         @Cached("target.createOperation()") TruffleObjectOperation operation) {
-            cached++;
+            cachedCount++;
             return operation.execute(frame, target, value);
         }
 
         @Specialization(replaces = "interopCached")
         protected Object interopGeneric(VirtualFrame frame, TruffleObject target, Object value) {
-            generic++;
+            genericCount++;
             return target.createOperation().execute(frame, target, value);
         }
     }